# openapi/ui-api.yaml
openapi: 3.1.0
info:
  title: rmcluster UI API
  version: 1.0.0
  summary: Backend contract required by the React frontend
  description: This spec defines the API endpoints required by the current frontend.
servers:
  - url: http://127.0.0.1:4917
    description: Local development
...
paths:
...
  /api/ui/connect-info:
    get:
      tags: [devices]
      summary: Get one-time device connect payload
      description: |
        Returns host, port, and single-use token for mobile device onboarding.
        Token should expire quickly (recommended <= 2 minutes) and be consumed once.
      operationId: getConnectInfo
      responses:
        '200':
          description: Connection payload for QR/code flow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectInfo'
        '500':
          $ref: '#/components/responses/Error'
...
  /api/ui/loading-status:
    get:
      tags: [ui]
      summary: Get current model loading phase and progress
      operationId: getLoadingStatus
      parameters:
        - name: model
          in: query
          required: false
          schema:
            type: string
          description: Selected chat model for loaded-device lookup
      responses:
        '200':
          description: Current loading state (empty model/phase when idle)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadingStatus'
...
  /api/ui/models/search:
    get:
      tags: [models]
      summary: Search Hugging Face model repos
      operationId: searchModels
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          description: User query or owner/repo substring
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '502':
          $ref: '#/components/responses/Error'
...
  /api/ui/models/local:
    post:
      tags: [models]
      summary: Upload local GGUF model
      operationId: uploadLocalModel
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [model_file]
              properties:
                model_file:
                  type: string
                  format: binary
                  description: .gguf file
                name:
                  type: string
                parameters:
                  type: string
                quantization:
                  type: string
      responses:
        '201':
          description: Uploaded model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
        '400':
          $ref: '#/components/responses/Error'
...
  /api/ui/chats:
    post:
      tags: [chat]
      summary: Record that a chat session started
      description: |
        Simple metadata endpoint for the UI to register conversation lifecycle.
        This does not replace `/v1/chat/completions`.
      operationId: startChatSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartChatRequest'
      responses:
        '201':
          description: Chat session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSession'
        '400':
          $ref: '#/components/responses/Error'
  /api/ui/chats/{chat_id}/events:
    post:
      tags: [chat]
      summary: Append sent/received/error event for a chat
      description: Auto-creates a minimal session if the chat_id is unknown.
      operationId: appendChatEvent
      parameters:
        - $ref: '#/components/parameters/ChatId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatEventRequest'
      responses:
        '202':
          description: Event accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusAccepted'
        '400':
          $ref: '#/components/responses/Error'
  /api/ui/chats/{chat_id}:
    get:
      tags: [chat]
      summary: Get chat metadata and events
      operationId: getChatSession
      parameters:
        - $ref: '#/components/parameters/ChatId'
      responses:
        '200':
          description: Chat session detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSessionDetail'
        '404':
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    deviceBearerToken:
      type: http
      scheme: bearer
      bearerFormat: one-time-device-token
  parameters:
...
    ChatId:
      name: chat_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    LoadingStatus:
      type: object
      required: [model, phase, progress, loaded_devices]
      properties:
        model:
          type: string
          description: Model currently loading; empty when idle
        phase:
          type: string
          enum: [starting, downloading, loading_model, warming_up, '']
          description: Current phase; empty when idle
        progress:
          type: number
          description: Download progress 0–100; meaningful only when phase=downloading
        layers_on_gpu:
          type: integer
          description: Layers currently offloaded to remote GPU devices
        node_count:
          type: integer
          description: Total connected nodes, used as a generic fallback
        loaded_devices:
          type: array
          items:
            $ref: '#/components/schemas/LoadedDevice'
...
    ConnectInfo:
      type: object
      required: [host, port, token, connect_uri, token_expires_in_seconds]
      properties:
        host:
          type: string
          example: 192.168.1.42
        port:
          type: integer
          minimum: 1
          maximum: 65535
          example: 4917
        token:
          type: string
          pattern: '^[a-f0-9]{32}$'
          example: d8fa1b262f174b06a5ee42a2f2d68f4a
        connect_uri:
          type: string
          description: Must follow the frontend URI format exactly.
          pattern: '^rmcluster://connect\\?url=.*&port=[0-9]+&token=[a-f0-9]{32}$'
          example: rmcluster://connect?url=192.168.1.42&port=4917&token=d8fa1b262f174b06a5ee42a2f2d68f4a
        token_expires_in_seconds:
          type: integer
          minimum: 1
          example: 120
...
    LoadedDevice:
      type: object
      required: [id, hardware_model]
      properties:
        id:
          type: string
        nickname:
          type: string
        hardware_model:
          type: string
...
    Model:
      type: object
      required: [model, display_name, source, link_href, link_label]
      properties:
        model:
          type: string
          example: hf:bartowski/Llama-3.2-3B-Instruct-GGUF
        display_name:
          type: string
        parameters:
          type: string
        architecture:
          type: string
        quantization:
          type: string
        source:
          type: string
        link_href:
          type: string
          format: uri
        link_label:
          type: string
...
    SearchResult:
      type: object
      required: [model, display_name, downloads, link_href]
      properties:
        model:
          type: string
          description: Usually `hf:owner/repo` format
        display_name:
          type: string
        downloads:
          type: integer
        link_href:
          type: string
          format: uri
    SearchResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
...
    StartChatRequest:
      type: object
      required: [chat_id, model]
      properties:
        chat_id:
          type: string
          description: Client-provided stable conversation ID
        model:
          type: string
        started_at:
          type: string
          format: date-time
        user_id:
          type: string
          description: Reserved for future auth/multi-user support
    ChatSession:
      type: object
      required: [chat_id, model, started_at, status]
      properties:
        chat_id:
          type: string
        model:
          type: string
        started_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, closed]
    ChatEventRequest:
      type: object
      required: [event_type, timestamp]
      properties:
        event_type:
          type: string
          enum:
            [
              message_sent,
              token_received,
              message_completed,
              stream_error,
              chat_closed,
            ]
        message_id:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          description: Full message for `message_sent` or `message_completed`
        token:
          type: string
          description: Single streamed token/chunk when `event_type=token_received`
        error:
          type: string
        timestamp:
          type: string
          format: date-time
    ChatEvent:
      allOf:
        - $ref: '#/components/schemas/ChatEventRequest'
        - type: object
          required: [sequence]
          properties:
            sequence:
              type: integer
              minimum: 1
    ChatSessionDetail:
      allOf:
        - $ref: '#/components/schemas/ChatSession'
        - type: object
          required: [events]
          properties:
            events:
              type: array
              items:
                $ref: '#/components/schemas/ChatEvent'
...
    StatusAccepted:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted]
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string